streamlit/streamlit2025/requirements.txt
numpy==1.24.4
pandas
seaborn
plotly-express